+++ packages/avatar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-ui/avatar
 
+## 1.0.12
+
+### Patch Changes
+
+- Add loading props
+
 ## 1.0.11
 
 ### Patch Changes